feat: Optionally leave generated things for debugging purposes. (#15)

+++ aws-greengrass-testing-features/aws-greengrass-testing-features-api/src/main/java/com/aws/greengrass/testing/features/FileSteps.java
@@ -72,8 +72,10 @@ public class FileSteps {
                             logFile, testContext.testResultsPath(), ie);
                 }
             });
-            // Remove the rest
-            platform.files().delete(testContext.installRoot());
+            if (!testContext.cleanupContext().persistInstalledSofware()) {
+                // Remove the rest
+                platform.files().delete(testContext.installRoot());
+            }
         }
     }
 }
